Add signed update delivery

+++ app/src/play/java/com/example/furiganakeyboard/update/AppUpdateWorker.kt
@@ -16,7 +16,10 @@ class AppUpdateWorker(
     override suspend fun doWork(): Result {
         val updateInfo = awaitUpdateInfo() ?: return Result.retry()
         if (updateInfo.updateAvailability() == UpdateAvailability.UPDATE_AVAILABLE) {
-            AppUpdateNotifications.showIfNew(applicationContext, updateInfo.availableVersionCode())
+            AppUpdateNotifications.showIfNew(
+                applicationContext,
+                updateInfo.availableVersionCode().toLong()
+            )
         }
         return Result.success()
     }
